+++ park_miller.c
@@ -8,17 +8,16 @@
 #define q (m / a)
 #define r (m % a)
 
-static long long int seed = 1;
+static int64_t seed = 1;
 
-long long int PMrand() {
+uint32_t PMrand() {
 	// X_n+1 = (48271 * X_n) mod (2^31 - 1)
 	// X = a (hi*q + lo) (mod m)
 	//   = hi*a*q + a*lo (mod a*q+r)
 	//   = -hi*r + a*lo (mod a*q+r)
-	
-	long long int hi = seed / q;
-	long long int lo = seed % q;
+	int64_t hi = seed / q;
+	int64_t lo = seed % q;
 	seed = a * lo - hi * r;
 	if (seed < 0) seed += m;
-	return seed;
+	return (uint32_t)seed;
 }
